+++ example/lib/main.dart
@@ -1,15 +1,19 @@
 import 'package:animated_tree_view/animated_tree_view.dart';
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 
 import '../utils/utils.dart';
 
 const showSnackBar = false;
+const expandChildrenOnReady = true;
 
 void main() {
-  runApp(MyApp());
+  runApp(const MyApp());
 }
 
 class MyApp extends StatelessWidget {
+  const MyApp({super.key});
+
   // This widget is the root of your application.
   @override
   Widget build(BuildContext context) {
@@ -19,23 +23,22 @@ class MyApp extends StatelessWidget {
         primarySwatch: Colors.blue,
         visualDensity: VisualDensity.adaptivePlatformDensity,
       ),
-      home: MyHomePage(title: 'Simple Animated Tree Demo'),
+      home: const MyHomePage(title: 'Simple Animated Tree Demo'),
     );
   }
 }
 
 class MyHomePage extends StatefulWidget {
-  MyHomePage({Key? key, required this.title}) : super(key: key);
+  const MyHomePage({Key? key, required this.title}) : super(key: key);
 
   final String title;
 
   @override
-  _MyHomePageState createState() => _MyHomePageState();
+  MyHomePageState createState() => MyHomePageState();
 }
 
-class _MyHomePageState extends State<MyHomePage> {
-  final GlobalKey<TreeViewState> _treeKey = GlobalKey<TreeViewState>();
-  late final controller = _treeKey.currentState?.controller;
+class MyHomePageState extends State<MyHomePage> {
+  TreeViewController? _controller;
 
   @override
   Widget build(BuildContext context) {
@@ -43,37 +46,48 @@ class _MyHomePageState extends State<MyHomePage> {
       appBar: AppBar(
         title: Text(widget.title),
       ),
-      floatingActionButton: FloatingActionButton.extended(
-        onPressed: () {
-          if (sampleTree.isExpanded) {
-            controller?.collapseNode(sampleTree);
-          } else {
-            controller?.expandAllChildren(sampleTree);
-          }
-          setState(() {});
+      floatingActionButton: ValueListenableBuilder<bool>(
+        valueListenable: sampleTree.expansionNotifier,
+        builder: (context, isExpanded, _) {
+          return FloatingActionButton.extended(
+            onPressed: () {
+              if (sampleTree.isExpanded) {
+                _controller?.collapseNode(sampleTree);
+              } else {
+                _controller?.expandAllChildren(sampleTree);
+              }
+            },
+            label: isExpanded
+                ? const Text("Collapse all")
+                : const Text("Expand all"),
+          );
         },
-        label:
-            sampleTree.isExpanded ? Text("Collapse all") : Text("Expand all"),
       ),
       body: TreeView.simple(
-        key: _treeKey,
         tree: sampleTree,
+        showRootNode: true,
         expansionIndicatorBuilder: (context, node) =>
             ChevronIndicator.rightDown(
           tree: node,
           color: Colors.blue[700],
           padding: const EdgeInsets.all(8),
         ),
-        indentation: Indentation(style: IndentStyle.squareJoint),
+        indentation: const Indentation(style: IndentStyle.squareJoint),
         onItemTap: (item) {
-          print("Item tapped: ${item.key}");
-          if (showSnackBar)
+          if (kDebugMode) print("Item tapped: ${item.key}");
+
+          if (showSnackBar) {
             ScaffoldMessenger.of(context).showSnackBar(
               SnackBar(
                 content: Text("Item tapped: ${item.key}"),
                 duration: const Duration(milliseconds: 750),
               ),
             );
+          }
+        },
+        onTreeReady: (controller) {
+          _controller = controller;
+          if (expandChildrenOnReady) controller.expandAllChildren(sampleTree);
         },
         builder: (context, node) => Card(
           color: colorMapper[node.level.clamp(0, colorMapper.length - 1)]!,