+++ test/tree_view/fakes/fake_indexed_tree_view_widget.dart
@@ -1,11 +1,15 @@
 import 'package:animated_tree_view/animated_tree_view.dart';
 import 'package:flutter/material.dart';
-import 'package:tuple/tuple.dart';
 
 class FakeStatelessIndexedTreeView<T> extends StatelessWidget {
   final IndexedTreeNode<T> tree;
+  final TreeReadyCallback<T, IndexedTreeNode<T>>? onTreeReady;
 
-  const FakeStatelessIndexedTreeView({super.key, required this.tree});
+  const FakeStatelessIndexedTreeView({
+    super.key,
+    required this.tree,
+    this.onTreeReady,
+  });
 
   @override
   Widget build(BuildContext context) {
@@ -15,6 +19,7 @@ class FakeStatelessIndexedTreeView<T> extends StatelessWidget {
           tree: tree,
           expansionBehavior: ExpansionBehavior.scrollToLastChild,
           showRootNode: true,
+          onTreeReady: onTreeReady,
           builder: (context, node) => ListTile(
             title: Text("Item ${node.level}-${node.key}"),
             subtitle: Text('Level ${node.level}'),
@@ -52,7 +57,7 @@ class FakeStatefulIndexedTreeViewState
       home: Material(
         child: Stack(children: [
           TreeView.indexed(
-            tree: testIndexedTrees[stateCount].item1,
+            tree: testIndexedTrees[stateCount].$1,
             expansionBehavior: ExpansionBehavior.scrollToLastChild,
             showRootNode: true,
             builder: (context, node) => ListTile(
@@ -71,15 +76,18 @@ class FakeStatefulIndexedTreeViewState
   }
 }
 
-late final testIndexedTrees = <Tuple2<IndexedTreeNode, List<IndexedTreeNode>>>[
-  Tuple2(defaultIndexedTree, []),
-  Tuple2(nodesAddedIndexedTree, []),
-  Tuple2(nodesRemovedIndexedTree, [
-    IndexedTreeNode(key: "0C"),
-    IndexedTreeNode(key: "0C1C"),
-    IndexedTreeNode(key: "0C1C2A"),
-    IndexedTreeNode(key: "0C1C2A3C"),
-  ]),
+late final testIndexedTrees = <(IndexedTreeNode, List<IndexedTreeNode>)>[
+  (defaultIndexedTree, []),
+  (nodesAddedIndexedTree, []),
+  (
+    nodesRemovedIndexedTree,
+    [
+      IndexedTreeNode(key: "0C"),
+      IndexedTreeNode(key: "0C1C"),
+      IndexedTreeNode(key: "0C1C2A"),
+      IndexedTreeNode(key: "0C1C2A3C"),
+    ]
+  ),
 ];
 
 IndexedTreeNode get defaultIndexedTree => IndexedTreeNode.root()