+++ test/tree_view/fakes/fake_tree_view_widget.dart
@@ -1,11 +1,15 @@
 import 'package:animated_tree_view/animated_tree_view.dart';
 import 'package:flutter/material.dart';
-import 'package:tuple/tuple.dart';
 
 class FakeStatelessTreeView<T> extends StatelessWidget {
   final TreeNode<T> tree;
+  final TreeReadyCallback<T, TreeNode<T>>? onTreeReady;
 
-  const FakeStatelessTreeView({super.key, required this.tree});
+  const FakeStatelessTreeView({
+    super.key,
+    required this.tree,
+    this.onTreeReady,
+  });
 
   @override
   Widget build(BuildContext context) {
@@ -13,8 +17,8 @@ class FakeStatelessTreeView<T> extends StatelessWidget {
       home: Material(
         child: TreeView.simple(
           tree: tree,
-          expansionBehavior: ExpansionBehavior.scrollToLastChild,
           showRootNode: true,
+          onTreeReady: onTreeReady,
           builder: (context, node) => ListTile(
             title: Text("Item ${node.level}-${node.key}"),
             subtitle: Text('Level ${node.level}'),
@@ -49,7 +53,7 @@ class FakeStatefulTreeViewState extends State<FakeStatefulTreeView> {
       home: Material(
         child: Stack(children: [
           TreeView.simple(
-            tree: testTrees[stateCount].item1,
+            tree: testTrees[stateCount].$1,
             expansionBehavior: ExpansionBehavior.scrollToLastChild,
             showRootNode: true,
             builder: (context, node) => ListTile(
@@ -68,15 +72,18 @@ class FakeStatefulTreeViewState extends State<FakeStatefulTreeView> {
   }
 }
 
-late final testTrees = <Tuple2<TreeNode, List<TreeNode>>>[
-  Tuple2(defaultTree, []),
-  Tuple2(nodesAddedTree, []),
-  Tuple2(nodesRemovedTree, [
-    TreeNode(key: "0C"),
-    TreeNode(key: "0C1C"),
-    TreeNode(key: "0C1C2A"),
-    TreeNode(key: "0C1C2A3C"),
-  ]),
+late final testTrees = <(TreeNode, List<TreeNode>)>[
+  (defaultTree, []),
+  (nodesAddedTree, []),
+  (
+    nodesRemovedTree,
+    [
+      TreeNode(key: "0C"),
+      TreeNode(key: "0C1C"),
+      TreeNode(key: "0C1C2A"),
+      TreeNode(key: "0C1C2A3C"),
+    ]
+  ),
 ];
 
 TreeNode get defaultTree => TreeNode.root()
@@ -93,6 +100,34 @@ TreeNode get defaultTree => TreeNode.root()
             TreeNode(key: "0C1C2A")
               ..addAll([
                 TreeNode(key: "0C1C2A3C"),
+              ]),
+          ]),
+      ]),
+  ]);
+
+TreeNode get longTree => TreeNode.root()
+  ..addAll([
+    TreeNode(key: "0A"),
+    TreeNode(key: "0B"),
+    TreeNode(key: "0C"),
+    TreeNode(key: "0D"),
+    TreeNode(key: "0E"),
+    TreeNode(key: "0F"),
+    TreeNode(key: "0G"),
+    TreeNode(key: "0H"),
+    TreeNode(key: "0I"),
+    TreeNode(key: "0J"),
+    TreeNode(key: "0K"),
+    TreeNode(key: "0L"),
+    TreeNode(key: "0M"),
+    TreeNode(key: "0N"),
+    TreeNode(key: "0Z")
+      ..addAll([
+        TreeNode(key: "0Z1Z")
+          ..addAll([
+            TreeNode(key: "0Z1Z2A")
+              ..addAll([
+                TreeNode(key: "0Z1Z2A3Z"),
               ]),
           ]),
       ]),